# examples/advanced/api/swagger/greeter.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Greeter API",
    "description": "API for greeting users",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/greeter/{name}": {
      "get": {
        "summary": "Send a greeting",
        "description": "Send a greeting to the specified name",
        "operationId": "SayHello",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Name to greet"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "$ref": "#/definitions/SayHelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "tags": [
          "Greeter"
        ]
      }
    }
  },
  "definitions": {
    "SayHelloResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Greeting message"
        },
        "timestamp": {
          "type": "string",
          "format": "int64",
          "description": "Timestamp of the greeting"
        }
      },
      "description": "Response from SayHello"
    },
    "Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Any"
          }
        }
      }
    },
    "Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    }
  }
} 